Hay un error, no se carga la ventana de agregar nuevo en doctor y paciente, pero los seeder ya son segun lo que esta ahi adentro

+++ app/Http/Controllers/DoctorController.php
@@ -12,84 +12,111 @@ class DoctorController extends Controller
 {
   //Lista de doctores
   public function index()
-  {           
-
-        $doctors = Doctor::all();
+  { 
 
 
+    if(Auth::Patient()) 
+    {
 
-    return view('hospital.doctor.indexDoctors', compact('doctors'));
+
+      $doctors = Doctor::all();
+      return view('hospital.doctor.indexDoctors', compact('doctors'));
+    } 
+    return view('admin');
+
+
   }
 
   //Crear doctores
   public function create()
   {
+
+    if(Auth::Office())
+    {
+
       $defaultImg= new Options();
-    $defaultImg = $defaultImg->UserDefault();
-    
-    $offices = Office::all(); 
-    $specialities = Speciality::all();
-    return view('hospital.doctor.createDoctor',compact('offices','specialities','defaultImg'));
+      $defaultImg = $defaultImg->UserDefault();
+      
+      $offices = Office::all(); 
+      $specialities = Speciality::all();
+      return view('hospital.doctor.createDoctor',compact('offices','specialities','defaultImg'));
+    }
+    return view('admin');
+
   }
 
   //Almacenar doctor
   public function store(Request $request)
   {
-    $this->validate($request, [
-      'name'=>'required',
-      'birthdate'=>'required',
-      'telephoneNumber'=>'required',
-      'turno'=>'required',
-      'sexo'=>'required',
-      'cedula'=>'required',
-      'especialidad'=>'required'
-    ]);
+    if(Auth::Office())
+    {
+
+      $this->validate($request, [
+        'name'=>'required',
+        'birthdate'=>'required',
+        'telephoneNumber'=>'required',
+        'turno'=>'required',
+        'sexo'=>'required',
+        'cedula'=>'required',
+        'especialidad'=>'required'
+      ]);
 
     //Crear médico
-    $doctor = new Doctor;
-    $doctor->name = $request->input('name');
-    $doctor->birthdate = $request->input('birthdate');
-    $doctor->telephoneNumber = $request->input('telephoneNumber');
-    $doctor->turno = $request->input('turno');
-    $doctor->sexo = $request->input('sexo');
-    $doctor->cedula = $request->input('cedula');
+      $doctor = new Doctor;
+      $doctor->name = $request->input('name');
+      $doctor->birthdate = $request->input('birthdate');
+      $doctor->telephoneNumber = $request->input('telephoneNumber');
+      $doctor->turno = $request->input('turno');
+      $doctor->sexo = $request->input('sexo');
+      $doctor->cedula = $request->input('cedula');
 
 
 
 
-    $especialidades = $request->input('especialidad');
-    for($i=0;$i<count($especialidades);$i++)
-    {
-      $speciality_id =$especialidades[$i];
+      $especialidades = $request->input('especialidad');
+      for($i=0;$i<count($especialidades);$i++)
+      {
+        $speciality_id =$especialidades[$i];
 
-      echo $speciality_id." ";
-     $doctor->specialities()->attach( $speciality_id);
+        echo $speciality_id." ";
+        $doctor->specialities()->attach( $speciality_id);
 
-    }
+      }
 
 
-    $doctor->office_id = $request->input('office_id');
-    $doctor->inTime = $request->input('inTime');
-    $doctor->onTime = $request->input('onTime');
+      $doctor->office_id = $request->input('office_id');
+      $doctor->inTime = $request->input('inTime');
+      $doctor->onTime = $request->input('onTime');
 
-    $doctor->save();
+      $doctor->save();
 
-    return redirect('/doctor')->with('success', '¡El médico ha sido agregado con éxito!');
+      return redirect('/doctor')->with('success', '¡El médico ha sido agregado con éxito!');
+    }    
+    return view('admin');
   }
 
   //Mostrar información de doctor
   public function show($id)
   {
 
+    if(Auth::Patient())
+    {
 
-    $doctor=Doctor::find($id);
 
 
-    
-    return view('hospital.doctor.showDoctor', compact('doctor'))
-        ->with('patients', $doctor->patients)
-        ->with('offices', $doctor->offices)
-        ->with('appointments', $doctor->appointments);
+
+      $doctor=Doctor::find($id);
+
+
+      
+      return view('hospital.doctor.showDoctor', compact('doctor'))
+      ->with('patients', $doctor->patients)
+      ->with('offices', $doctor->offices)
+      ->with('appointments', $doctor->appointments);
+    }
+
+    return view('admin');
+
   }
 
   //Actualizar doctor
@@ -106,10 +133,10 @@ class DoctorController extends Controller
         return view('admin');
       }
 
-    $offices = Office::all();
-    $specialities = Speciality::all();
-    $doctor= Doctor::find($id);
-    return view('hospital.doctor.editDoctor', compact('doctor','specialities','offices'));
+      $offices = Office::all();
+      $specialities = Speciality::all();
+      $doctor= Doctor::find($id);
+      return view('hospital.doctor.editDoctor', compact('doctor','specialities','offices'));
     }
 
     return view('admin');
@@ -118,57 +145,74 @@ class DoctorController extends Controller
   //Método update
   public function update(Request $request)
   {
-    $this->validate($request, [
-      'name'=>'required',
-      'doctor_id',
-      'birthdate'=>'required',
-      'telephoneNumber'=>'required',
-      'turno'=>'required',
-      'sexo'=>'required',
-      'cedula'=>'required',
-      'especialidad'=>'required'
-    ]);
 
-
-
-    $doctor = Doctor::find($request->input('doctor_id'));
-
-    $doctor->specialities()->detach();
-
-
-    $especialidades = $request->input('especialidad');
-    for($i=0;$i<count($especialidades);$i++)
+    if(Auth::Doctor())
     {
-      $speciality_id =$especialidades[$i];
 
-      echo $speciality_id." ";
-     $doctor->specialities()->attach( $speciality_id);
 
-    }
+
+      $this->validate($request, [
+        'name'=>'required',
+        'doctor_id',
+        'birthdate'=>'required',
+        'telephoneNumber'=>'required',
+        'turno'=>'required',
+        'sexo'=>'required',
+        'cedula'=>'required',
+        'especialidad'=>'required'
+      ]);
+
+
+
+      $doctor = Doctor::find($request->input('doctor_id'));
+
+      $doctor->specialities()->detach();
+
+
+      $especialidades = $request->input('especialidad');
+      for($i=0;$i<count($especialidades);$i++)
+      {
+        $speciality_id =$especialidades[$i];
+
+        echo $speciality_id." ";
+        $doctor->specialities()->attach( $speciality_id);
+
+      }
 
 
 
     //Editar médico
-    $doctor->name = $request->input('name');
-    $doctor->birthdate = $request->input('birthdate');
-    $doctor->telephoneNumber = $request->input('telephoneNumber');
-    $doctor->turno = $request->input('turno');
-    $doctor->sexo = $request->input('sexo');
-    $doctor->cedula = $request->input('cedula');
+      $doctor->name = $request->input('name');
+      $doctor->birthdate = $request->input('birthdate');
+      $doctor->telephoneNumber = $request->input('telephoneNumber');
+      $doctor->turno = $request->input('turno');
+      $doctor->sexo = $request->input('sexo');
+      $doctor->cedula = $request->input('cedula');
 
-     $doctor->office_id = $request->input('office_id');
-    $doctor->inTime = $request->input('inTime');
-    $doctor->outTime = $request->input('outTime');
+      $doctor->office_id = $request->input('office_id');
+      $doctor->inTime = $request->input('inTime');
+      $doctor->outTime = $request->input('outTime');
 
-    $doctor->save();
+      $doctor->save();
 
-    return redirect('/doctor/'.$doctor->id)->with('success','¡El médico ha sido actualizado con éxito!');
+      return redirect('/doctor/'.$doctor->id)->with('success','¡El médico ha sido actualizado con éxito!');
+
+    }
+    return view('admin');
+
   }
 
   //Eliminar doctor
   public function destroy(Doctor $doctor)
   {
-    $doctor->delete();
-    return redirect('/doctor')->with('success', '¡El médico ha sido eliminado con éxito!');
+    if(Auth::Office())
+    {
+
+      $doctor->delete();
+      return redirect('/doctor')->with('success', '¡El médico ha sido eliminado con éxito!');
+    } 
+    return view('admin');
+
+
   }
 }
